divided main js into modules
